requirements.txt
Flask==2.3.3
SQLAlchemy==2.0.20
Flask-SQLAlchemy==3.1.1
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0